글쓰기 페이지 카테고리 서버 연동
