+++ frontend/pages/Write/CategorySelector/index.js
@@ -1,9 +1,10 @@
-import $ from '../../util/domControll';
-import ElementBuilder from '../../lib/ElementBuilder';
+import $ from '../../../util/domControll';
+import ElementBuilder from '../../../lib/ElementBuilder';
 import CategorySelectContainer from './CategorySelectorContainer';
 
 export default class CategorySelector extends ElementBuilder {
   constructElement() {
+    const { categories } = this.props;
     const $element = $.create('div').addClass('category-selector--container');
     if (this.props.invisible) {
       $element.addClass('invisible');
@@ -17,7 +18,9 @@ export default class CategorySelector extends ElementBuilder {
       parent: this,
       buttonState: this.props.buttonState,
       setButtonState: this.props.setButtonState,
+      categories,
     });
+
     return $element;
   }
 }
